test: expand E2E suite and fix 5 test program bugs
Expanded E2E test suite with inline run tests for simple-sum and odd-even
(Level 5a/5b), restructuring Level 5 integration examples to Level 5c and
mandelbrot promotion to Level 5d. This improves test granularity and enables
earlier detection of varying type regressions.

Fixed 5 program bugs preventing SPMD compilation:
- union-type-generics: shift operand type mismatch ([]int → []uint16)
- pointer-varying: lanes.From(array) requires slice syntax (array[:])
- base64-decoder: constrained range[N] missing upper bound (range[4] → range[4] 4)
- select-with-varying-channels: replaced invalid "go for {}" with regular "for {}"
- map-restrictions: varying-to-uniform struct literal via reduce.From() conversion

All 5 fixes synchronized across examples/ and test/integration/spmd/.

E2E results improved: 44→46 total tests, 9→11 run pass, 16→19 compile pass,
17→16 compile fail (reduced from 18 in prior commit).

+++ examples/union-type-generics/main.go
@@ -147,7 +147,7 @@ func demonstrateShiftOperations() {
 	// Test with constrained varying
 	fmt.Println("\n--- Constrained Shift Operations ---")
 	constrainedValues := lanes.From([]uint16{0x00FF, 0x0F0F, 0xF0F0, 0xFF00})
-	constrainedShifts := lanes.From([]int{1, 2, 3, 4})
+	constrainedShifts := lanes.From([]uint16{1, 2, 3, 4})
 
 	constrainedLeft := lanes.ShiftLeft(constrainedValues, constrainedShifts)
 	constrainedRight := lanes.ShiftRight(constrainedValues, constrainedShifts)
